Enum meggl::BitmapRefMut

pub enum BitmapRefMut<'a> {
    Indexed(BitmapRefMut8<'a>),
    Argb32(BitmapRefMut32<'a>),
}

Variants

Indexed(BitmapRefMut8<'a>)

Argb32(BitmapRefMut32<'a>)

Implementations

impl<'a> BitmapRefMut<'a>

pub fn as_const(&'a self) -> BitmapRef<'a>

pub fn view(&mut self, rect: Rect) -> Option<Self>

Returns a subset of the specified range of bitmaps. The function returns None if the rectangle points outside the range of the bitmap.

impl BitmapRefMut<'_>

pub fn copy(&mut self, origin: Point, rect: Rect)

pub fn blt_transparent( &mut self, src: &BitmapRef<'_>, origin: Point, rect: Rect, color_key: IndexedColor )

pub fn map_indexed<F, R>(&mut self, f: F) -> Option<R>
where F: FnOnce(&mut BitmapRefMut8<'_>) -> R,

pub fn map_argb32<F, R>(&mut self, f: F) -> Option<R>
where F: FnOnce(&mut BitmapRefMut32<'_>) -> R,

pub const fn color_mode(&self) -> usize