Struct meggl::OperationalBitmap

pub struct OperationalBitmap { /* private fields */ }

A special bitmap type that can be used for operations such as transparency and shading.

Implementations

impl OperationalBitmap

pub const fn new(size: Size) -> Self

pub fn from_slice(slice: &[u8], size: Size) -> Self

pub fn from_vec(vec: Vec<u8>, size: Size) -> Self

pub fn from_pixels<F, T>(data: &T, kernel: F) -> OperationalBitmap
where T: GetPixel + ?Sized, F: FnMut(<T as Drawable>::ColorType) -> u8,

pub fn reset(&mut self)

pub fn fill(&mut self, color: u8)

pub fn draw_line<F>(&mut self, c1: Point, c2: Point, kernel: F)
where F: FnMut(&mut OperationalBitmap, Point),

Draws a straight line at high speed using Bresenhamโs line algorithm.

pub fn draw_line_anti_aliasing_f<F>(&mut self, c1: Point, c2: Point, kernel: F)
where F: FnMut(&mut OperationalBitmap, Point, f64),

Draws an anti-aliased line using Xiaolin Wuโs algorithm.

pub fn draw_line_anti_aliasing_i<F>( &mut self, c1: Point, c2: Point, scale: isize, kernel: F )
where F: FnMut(&mut OperationalBitmap, Point, u8),

๐Deprecated

pub fn blur(&mut self, radius: isize, level: usize)

Like box linear filter

pub fn blt_to<T, F>(&self, dest: &mut T, origin: Point, rect: Rect, kernel: F)
where T: Drawable + GetPixel + SetPixel, F: FnMut(u8, <T as Drawable>::ColorType) -> <T as Drawable>::ColorType,

pub fn blt_shadow( &self, dest: &mut BitmapRefMut32<'_>, origin: Point, rect: Rect )

pub fn blt_from<T, F>(&mut self, src: &T, origin: Point, rect: Rect, kernel: F)
where T: GetPixel, F: FnMut(<T as Drawable>::ColorType, u8) -> u8,

pub fn draw_to( &self, dest: &mut BitmapRefMut<'_>, origin: Point, rect: Rect, color: Color )