Trait meggl::GetPixel

pub trait GetPixel: Drawable {
    // Required method
    unsafe fn get_pixel_unchecked(&self, point: Point) -> Self::ColorType;

    // Provided methods
    fn get_pixel(&self, point: Point) -> Option<Self::ColorType> { ... }
    fn to_operational<F>(&self, kernel: F) -> OperationalBitmap
       where F: FnMut(<Self as Drawable>::ColorType) -> u8 { ... }
    fn all_pixels<'a>(&'a self) -> AllPixels<'a, Self>  { ... }
}

Required Methods

unsafe fn get_pixel_unchecked(&self, point: Point) -> Self::ColorType

Safety

The point must be within the size range.

Provided Methods

fn get_pixel(&self, point: Point) -> Option<Self::ColorType>

fn to_operational<F>(&self, kernel: F) -> OperationalBitmap
where F: FnMut(<Self as Drawable>::ColorType) -> u8,

fn all_pixels<'a>(&'a self) -> AllPixels<'a, Self>

Returns an iterator that enumerates all pixels of the bitmap in order from left to right and top to bottom.

Object Safety

This trait is not object safe.