Trait meggl::PixelColor

pub trait PixelColor: Sized + Copy + Clone + PartialEq + Eq + Default {
    // Provided method
    fn stride_for(width: isize) -> usize { ... }
}

Common color trait

Provided Methods

fn stride_for(width: isize) -> usize

This value is used to calculate the address of a raster image that supports this color format.

Object Safety

This trait is not object safe.